functional_programming/dissaster_tracker/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.1
celery==5.3.4
redis==5.0.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
geopy==2.4.0
python-dotenv==1.0.0